Fix bug
Fixed an issue where if you deleted a book, all subsequent book buttons
would break. This was caused from using the current myLibrary list length
at button creation as an identifier, so when a book was deleted,
the list index changed and the button called the wrong object in
the list. Fixed the read button by adding a data attribute to the button.
Fixed delete button by passing the object as an argument in the createCard
function and calling the method directly from the object.